+++ scalr/resource_scalr_workspace_migrate.go
@@ -113,6 +113,11 @@ func resourceScalrWorkspaceResourceV0() *schema.Resource {
 }
 
 func resourceScalrWorkspaceStateUpgradeV0(rawState map[string]interface{}, meta interface{}) (map[string]interface{}, error) {
+	if _, ok := rawState["external_id"]; !ok {
+		// Due to migration drift, schema-versionV0 can already contain 'id' field
+		// so we can skip V0->V1 the migration.
+		return rawState, nil
+	}
 	rawState["id"] = rawState["external_id"]
 	delete(rawState, "external_id")
 	delete(rawState, "ssh_key_id")
@@ -212,10 +217,14 @@ func resourceScalrWorkspaceResourceV1() *schema.Resource {
 
 func resourceScalrWorkspaceStateUpgradeV1(rawState map[string]interface{}, meta interface{}) (map[string]interface{}, error) {
 	if rawState["vcs_repo"] != nil {
-		vcsRepo := rawState["vcs_repo"].(map[string]interface{})
+		vcsRepos := rawState["vcs_repo"].([]interface{})
+		if len(vcsRepos) == 0 {
+			return rawState, nil
+		}
+		vcsRepo := vcsRepos[0].(map[string]interface{})
 		rawState["vcs_provider_id"] = vcsRepo["oauth_token_id"]
 		delete(vcsRepo, "oauth_token_id")
-		rawState["vcs_repo"] = vcsRepo
+		rawState["vcs_repo"] = []interface{}{vcsRepo}
 	}
 	return rawState, nil
 }